requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pytest==7.4.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
